requirements.txt
termcolor==2.2.0
PyYAML==6.0